# Remove restrictions on Label types