--- a/crates/bevy_derive/src/lib.rs
+++ b/crates/bevy_derive/src/lib.rs
@@ -94,6 +94,8 @@ pub fn derive_enum_variant_meta(input: TokenStream) -> TokenStream {
 pub fn derive_app_label(input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as syn::DeriveInput);
     let mut trait_path = BevyManifest::default().get_path("bevy_app");
-    trait_path.segments.push(format_ident!("AppLabel").into());
+    trait_path
+        .segments
+        .push(format_ident!("IntoAppLabel").into());
     derive_label(input, &trait_path, "app_label")
 }