… crates/bevy_ecs/examples/derive_label.rs
… crates/bevy_utils/Cargo.toml
… crates/bevy_utils/src/intern.rs
… crates/bevy_utils/src/label.rs
… crates/bevy_utils/src/lib.rs